requirements.txt
django-jalali
whitenoise
